added routes, integrated nav bar
