+++ client/src/pages/Auth/Auth.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Auth = () => {
-	return <div></div>;
+	return <div>Auth</div>;
 };
 
 export default Auth;
