+++ client/src/pages/Scores/Scores.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Scores = () => {
-	return <div></div>;
+	return <div>Scores</div>;
 };
 
 export default Scores;
