+++ package.json
@@ -13,6 +13,7 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@types/react-router-dom": "^5.3.2",
 		"dotenv": "^10.0.0",
 		"express": "^4.17.2",
 		"mongodb": "^4.2.2",
